make OICondition inner class of OIFragment
